feat(core): fixed some code smells and bugs of sonar
