+++ libs/core/src/main/java/com/sourcefuse/jarc/core/CoreApplication.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 @EnableJpaAuditing
 public class CoreApplication {
 
-  public static void main(String[] args) {
-    SpringApplication.run(CoreApplication.class, args);
-  }
+    public static void main(String[] args) {
+        SpringApplication.run(CoreApplication.class, args);
+    }
 }
