+++ libs/core/src/main/java/com/sourcefuse/jarc/core/adapters/LocalDateTimeTypeAdapter.java
@@ -1,0 +1,35 @@
+package com.sourcefuse.jarc.core.adapters;
+
+import com.google.gson.TypeAdapter;
+import com.google.gson.stream.JsonReader;
+import com.google.gson.stream.JsonToken;
+import com.google.gson.stream.JsonWriter;
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class LocalDateTimeTypeAdapter extends TypeAdapter<LocalDateTime> {
+
+    private static final DateTimeFormatter FORMATTER =
+        DateTimeFormatter.ISO_LOCAL_DATE_TIME;
+
+    @Override
+    public void write(JsonWriter out, LocalDateTime value) throws IOException {
+        if (value == null) {
+            out.nullValue();
+        } else {
+            out.value(value.format(FORMATTER));
+        }
+    }
+
+    @Override
+    public LocalDateTime read(JsonReader in) throws IOException {
+        if (in.peek() == JsonToken.NULL) {
+            in.nextNull();
+            return null;
+        } else {
+            String dateString = in.nextString();
+            return LocalDateTime.parse(dateString, FORMATTER);
+        }
+    }
+}
