+++ libs/core/src/main/java/com/sourcefuse/jarc/core/awares/ApplicationAwareBeanUtils.java
@@ -9,26 +9,20 @@ import org.springframework.stereotype.Service;
 @Service
 public class ApplicationAwareBeanUtils implements ApplicationContextAware {
 
-  private static ApplicationContext context;
+    private static ApplicationContext context;
 
-  @Override
-  public void setApplicationContext(ApplicationContext applicationContext)
-    throws BeansException {
-    context = applicationContext;
-  }
-
-  public static <T> T getBean(Class<T> beanClass) {
-    return context.getBean(beanClass);
-  }
-
-  public static EntityManager getNewEntityManager()
-    throws IllegalArgumentException {
-    EntityManager em = context.getBean(EntityManager.class);
-    if (em == null) {
-      throw new IllegalArgumentException(
-        "Entity Manager is null can not proceed to save audit log"
-      );
+    @Override
+    public void setApplicationContext(ApplicationContext applicationContext)
+        throws BeansException {
+        context = applicationContext;
     }
-    return em.getEntityManagerFactory().createEntityManager();
-  }
+
+    public static <T> T getBean(Class<T> beanClass) {
+        return context.getBean(beanClass);
+    }
+
+    public static EntityManager getNewEntityManager() {
+        EntityManager em = context.getBean(EntityManager.class);
+        return em.getEntityManagerFactory().createEntityManager();
+    }
 }
