+++ libs/core/src/main/java/com/sourcefuse/jarc/core/constants/AuditActions.java
@@ -1,22 +1,22 @@
 package com.sourcefuse.jarc.core.constants;
 
 public enum AuditActions {
-  SAVE("SAVE"),
-  UPDATE("UPDATE"),
-  DELETE("DELETE");
+    SAVE("SAVE"),
+    UPDATE("UPDATE"),
+    DELETE("DELETE");
 
-  private final String name;
+    private final String name;
 
-  AuditActions(String value) {
-    this.name = value;
-  }
+    AuditActions(String value) {
+        this.name = value;
+    }
 
-  public String value() {
-    return this.name;
-  }
+    public String value() {
+        return this.name;
+    }
 
-  @Override
-  public String toString() {
-    return name;
-  }
+    @Override
+    public String toString() {
+        return name;
+    }
 }
