+++ libs/core/src/main/java/com/sourcefuse/jarc/core/models/base/SoftDeleteEntity.java
@@ -4,7 +4,7 @@ import com.sourcefuse.jarc.core.entitylisteners.SoftDeleteEntityListner;
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
-import java.util.Date;
+import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -17,10 +17,10 @@ import lombok.NoArgsConstructor;
 @EntityListeners(SoftDeleteEntityListner.class)
 public abstract class SoftDeleteEntity implements BaseEntity {
 
-  @Column(columnDefinition = "boolean default false")
-  boolean deleted;
+    @Column(columnDefinition = "boolean default false")
+    boolean deleted;
 
-  Date deletedOn;
+    LocalDateTime deletedOn;
 
-  UUID deletedBy;
+    UUID deletedBy;
 }
