+++ libs/core/src/main/java/com/sourcefuse/jarc/core/models/session/CurrentUser.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 @NoArgsConstructor
 public class CurrentUser<T extends BaseEntity> {
 
-  private T user;
+    private T user;
 }
