+++ libs/core/src/test/java/com/sourcefuse/jarc/core/CoreApplicationTests.java
@@ -1,6 +1,6 @@
 package com.sourcefuse.jarc.core;
 
-import com.sourcefuse.jarc.core.softdelete.SoftDeletesRepositoryImpl;
+import com.sourcefuse.jarc.core.repositories.SoftDeletesRepositoryImpl;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
@@ -9,6 +9,6 @@ import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 @EnableJpaRepositories(repositoryBaseClass = SoftDeletesRepositoryImpl.class)
 class CoreApplicationTests {
 
-  @Test
-  void contextLoads() {}
+    @Test
+    void contextLoads() {}
 }
