+++ libs/core/src/test/java/com/sourcefuse/jarc/core/test/models/Role.java
@@ -25,11 +25,11 @@ import lombok.ToString;
 @EntityListeners(AuditLogEntityListener.class)
 public class Role extends SoftDeleteEntity {
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.UUID)
-  private UUID id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
 
-  String name;
+    String name;
 
-  String permissions;
+    String permissions;
 }
