+++ libs/core/src/test/java/com/sourcefuse/jarc/core/test/models/User.java
@@ -21,13 +21,13 @@ import lombok.Setter;
 @Setter
 public class User implements BaseEntity {
 
-  @Id
-  @GeneratedValue(strategy = GenerationType.UUID)
-  private UUID id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.UUID)
+    private UUID id;
 
-  private String firstName;
-  private String lastName;
-  private String middleName;
-  private String username;
-  private String email;
+    private String firstName;
+    private String lastName;
+    private String middleName;
+    private String username;
+    private String email;
 }
